requirements.txt
openai
pandas
langchain
faiss-cpu
python-dotenv
streamlit
kagglehub
ipykernel
chromadb
langchain_community
tiktoken
gradio
langchain_openai